requirements.txt
absl-py==0.13.0
argon2-cffi==21.1.0
async-generator==1.10
attrs==21.2.0
awscli==1.20.37
backcall==0.2.0
bleach==4.1.0
boto3==1.18.37
botocore==1.21.37
cachetools==4.2.2
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.3
configparser==5.0.2
cssselect2==0.4.1
cycler==0.10.0
dataclasses==0.6
decorator==5.0.9
defusedxml==0.7.1
docker-pycreds==0.4.0
docopt==0.6.2
docstring-to-markdown==0.9
docutils==0.15.2
entrypoints==0.3
future==0.18.2
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.35.0
google-auth-oauthlib==0.4.6
greenlet==1.1.1
grpcio==1.40.0
idna==3.2
importlib-metadata==3.10.1
ipdb==0.13.9
ipykernel==5.5.5
ipython>=7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.4
jedi==0.18.0
jedi-language-server==0.34.3
Jinja2==3.0.1
jmespath==0.10.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.2
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.1
kaleido==0.2.1
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
msgpack==1.0.2
nbclient==0.5.4
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.3
numpy==1.19.5
oauthlib==3.1.1
opencv-python-headless==4.5.3.56
packaging==21.0
pandocfilters==1.4.3
parso==0.8.2
path==16.2.0
path.py==12.5.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
plotly==5.3.1
pnglatex==1.1
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.20
protobuf==3.17.3
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
pygls==0.11.2
Pygments==2.10.0
pynvim==0.4.3
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-xlib==0.31
PyYAML==5.4.1
pyzmq==22.2.1
qtconsole==5.1.1
QtPy==1.11.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.5.0
Send2Trash==1.8.0
sentry-sdk==1.3.1
shortuuid==1.0.1
six==1.16.0
smmap==4.0.0
subprocess32==3.5.4
tenacity==8.0.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
tinycss2==1.1.0
toml==0.10.2
torch==1.6.0
torchvision==0.7.0
tornado==6.1
tqdm==4.62.2
traitlets==4.3.3
typeguard==2.12.1
typing-extensions==3.10.0.2
ueberzug==18.1.9
urllib3==1.26.6
wandb==0.12.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
yacs==0.1.8
zipp==3.5.0
